+++ lib/maxmind/geoip2/model/anonymous_plus.rb
@@ -27,6 +27,7 @@ module MaxMind
           date_string = get('network_last_seen')
 
           if !date_string
+            @network_last_seen = nil
             return nil
           end